+++ CA2/server/serverCore.cpp
@@ -339,7 +339,7 @@ bool Client_Like_ConnectionFinderByPort::operator()(Client_Like_Connection *cl)
 }
 
 void clear_buff(char *x,size_t s){
-	for(int i=0;i<s;i++){
+	for(int i=0;i<(int)s;i++){
 		x[i] = 0;
 	}
 }
